Move testing cli tests into test_testing.py.
